requirements.txt
streamlit==1.31.1
google-generativeai==0.3.2
PyPDF2==3.0.1
Pillow==10.2.0